refactor: drop store barrel file
