+++ packages/app/src/api/graphql-env.d.ts
@@ -766,7 +766,7 @@ export type introspection_types = {
     'OrgUpdateMemberAuditEntry': { kind: 'OBJECT'; name: 'OrgUpdateMemberAuditEntry'; fields: { 'action': { name: 'action'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'actor': { name: 'actor'; type: { kind: 'UNION'; name: 'AuditEntryActor'; ofType: null; } }; 'actorIp': { name: 'actorIp'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'actorLocation': { name: 'actorLocation'; type: { kind: 'OBJECT'; name: 'ActorLocation'; ofType: null; } }; 'actorLogin': { name: 'actorLogin'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'actorResourcePath': { name: 'actorResourcePath'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; 'actorUrl': { name: 'actorUrl'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; 'createdAt': { name: 'createdAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'PreciseDateTime'; ofType: null; }; } }; 'id': { name: 'id'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'ID'; ofType: null; }; } }; 'operationType': { name: 'operationType'; type: { kind: 'ENUM'; name: 'OperationType'; ofType: null; } }; 'organization': { name: 'organization'; type: { kind: 'OBJECT'; name: 'Organization'; ofType: null; } }; 'organizationName': { name: 'organizationName'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'organizationResourcePath': { name: 'organizationResourcePath'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; 'organizationUrl': { name: 'organizationUrl'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; 'permission': { name: 'permission'; type: { kind: 'ENUM'; name: 'OrgUpdateMemberAuditEntryPermission'; ofType: null; } }; 'permissionWas': { name: 'permissionWas'; type: { kind: 'ENUM'; name: 'OrgUpdateMemberAuditEntryPermission'; ofType: null; } }; 'user': { name: 'user'; type: { kind: 'OBJECT'; name: 'User'; ofType: null; } }; 'userLogin': { name: 'userLogin'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'userResourcePath': { name: 'userResourcePath'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; 'userUrl': { name: 'userUrl'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; }; };
     'OrgUpdateMemberAuditEntryPermission': { name: 'OrgUpdateMemberAuditEntryPermission'; enumValues: 'ADMIN' | 'READ'; };
     'OrgUpdateMemberRepositoryCreationPermissionAuditEntry': { kind: 'OBJECT'; name: 'OrgUpdateMemberRepositoryCreationPermissionAuditEntry'; fields: { 'action': { name: 'action'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'actor': { name: 'actor'; type: { kind: 'UNION'; name: 'AuditEntryActor'; ofType: null; } }; 'actorIp': { name: 'actorIp'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'actorLocation': { name: 'actorLocation'; type: { kind: 'OBJECT'; name: 'ActorLocation'; ofType: null; } }; 'actorLogin': { name: 'actorLogin'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'actorResourcePath': { name: 'actorResourcePath'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; 'actorUrl': { name: 'actorUrl'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; 'canCreateRepositories': { name: 'canCreateRepositories'; type: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; } }; 'createdAt': { name: 'createdAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'PreciseDateTime'; ofType: null; }; } }; 'id': { name: 'id'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'ID'; ofType: null; }; } }; 'operationType': { name: 'operationType'; type: { kind: 'ENUM'; name: 'OperationType'; ofType: null; } }; 'organization': { name: 'organization'; type: { kind: 'OBJECT'; name: 'Organization'; ofType: null; } }; 'organizationName': { name: 'organizationName'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'organizationResourcePath': { name: 'organizationResourcePath'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; 'organizationUrl': { name: 'organizationUrl'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; 'user': { name: 'user'; type: { kind: 'OBJECT'; name: 'User'; ofType: null; } }; 'userLogin': { name: 'userLogin'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'userResourcePath': { name: 'userResourcePath'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; 'userUrl': { name: 'userUrl'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; 'visibility': { name: 'visibility'; type: { kind: 'ENUM'; name: 'OrgUpdateMemberRepositoryCreationPermissionAuditEntryVisibility'; ofType: null; } }; }; };
-    'OrgUpdateMemberRepositoryCreationPermissionAuditEntryVisibility': { name: 'OrgUpdateMemberRepositoryCreationPermissionAuditEntryVisibility'; enumValues: 'ALL' | 'INTERNAL' | 'NONE' | 'PRIVATE' | 'PRIVATE_INTERNAL' | 'PUBLIC' | 'VITE_INTERNAL' | 'VITE_PRIVATE'; };
+    'OrgUpdateMemberRepositoryCreationPermissionAuditEntryVisibility': { name: 'OrgUpdateMemberRepositoryCreationPermissionAuditEntryVisibility'; enumValues: 'ALL' | 'INTERNAL' | 'NONE' | 'PRIVATE' | 'PRIVATE_INTERNAL' | 'PUBLIC' | 'PUBLIC_INTERNAL' | 'PUBLIC_PRIVATE'; };
     'OrgUpdateMemberRepositoryInvitationPermissionAuditEntry': { kind: 'OBJECT'; name: 'OrgUpdateMemberRepositoryInvitationPermissionAuditEntry'; fields: { 'action': { name: 'action'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'actor': { name: 'actor'; type: { kind: 'UNION'; name: 'AuditEntryActor'; ofType: null; } }; 'actorIp': { name: 'actorIp'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'actorLocation': { name: 'actorLocation'; type: { kind: 'OBJECT'; name: 'ActorLocation'; ofType: null; } }; 'actorLogin': { name: 'actorLogin'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'actorResourcePath': { name: 'actorResourcePath'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; 'actorUrl': { name: 'actorUrl'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; 'canInviteOutsideCollaboratorsToRepositories': { name: 'canInviteOutsideCollaboratorsToRepositories'; type: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; } }; 'createdAt': { name: 'createdAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'PreciseDateTime'; ofType: null; }; } }; 'id': { name: 'id'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'ID'; ofType: null; }; } }; 'operationType': { name: 'operationType'; type: { kind: 'ENUM'; name: 'OperationType'; ofType: null; } }; 'organization': { name: 'organization'; type: { kind: 'OBJECT'; name: 'Organization'; ofType: null; } }; 'organizationName': { name: 'organizationName'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'organizationResourcePath': { name: 'organizationResourcePath'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; 'organizationUrl': { name: 'organizationUrl'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; 'user': { name: 'user'; type: { kind: 'OBJECT'; name: 'User'; ofType: null; } }; 'userLogin': { name: 'userLogin'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'userResourcePath': { name: 'userResourcePath'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; 'userUrl': { name: 'userUrl'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; }; };
     'Organization': { kind: 'OBJECT'; name: 'Organization'; fields: { 'announcement': { name: 'announcement'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'announcementBanner': { name: 'announcementBanner'; type: { kind: 'OBJECT'; name: 'AnnouncementBanner'; ofType: null; } }; 'announcementCreatedAt': { name: 'announcementCreatedAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; 'announcementExpiresAt': { name: 'announcementExpiresAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; 'announcementUserDismissible': { name: 'announcementUserDismissible'; type: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; } }; 'anyPinnableItems': { name: 'anyPinnableItems'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'archivedAt': { name: 'archivedAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; 'auditLog': { name: 'auditLog'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'OrganizationAuditEntryConnection'; ofType: null; }; } }; 'avatarUrl': { name: 'avatarUrl'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'URI'; ofType: null; }; } }; 'createdAt': { name: 'createdAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; 'databaseId': { name: 'databaseId'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'description': { name: 'description'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'descriptionHTML': { name: 'descriptionHTML'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'domains': { name: 'domains'; type: { kind: 'OBJECT'; name: 'VerifiableDomainConnection'; ofType: null; } }; 'email': { name: 'email'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'enterpriseOwners': { name: 'enterpriseOwners'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'OrganizationEnterpriseOwnerConnection'; ofType: null; }; } }; 'estimatedNextSponsorsPayoutInCents': { name: 'estimatedNextSponsorsPayoutInCents'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Int'; ofType: null; }; } }; 'hasSponsorsListing': { name: 'hasSponsorsListing'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'id': { name: 'id'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'ID'; ofType: null; }; } }; 'interactionAbility': { name: 'interactionAbility'; type: { kind: 'OBJECT'; name: 'RepositoryInteractionAbility'; ofType: null; } }; 'ipAllowListEnabledSetting': { name: 'ipAllowListEnabledSetting'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'ENUM'; name: 'IpAllowListEnabledSettingValue'; ofType: null; }; } }; 'ipAllowListEntries': { name: 'ipAllowListEntries'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'IpAllowListEntryConnection'; ofType: null; }; } }; 'ipAllowListForInstalledAppsEnabledSetting': { name: 'ipAllowListForInstalledAppsEnabledSetting'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'ENUM'; name: 'IpAllowListForInstalledAppsEnabledSettingValue'; ofType: null; }; } }; 'isSponsoredBy': { name: 'isSponsoredBy'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'isSponsoringViewer': { name: 'isSponsoringViewer'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'isVerified': { name: 'isVerified'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'itemShowcase': { name: 'itemShowcase'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'ProfileItemShowcase'; ofType: null; }; } }; 'lifetimeReceivedSponsorshipValues': { name: 'lifetimeReceivedSponsorshipValues'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'SponsorAndLifetimeValueConnection'; ofType: null; }; } }; 'location': { name: 'location'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'login': { name: 'login'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'mannequins': { name: 'mannequins'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'MannequinConnection'; ofType: null; }; } }; 'memberStatuses': { name: 'memberStatuses'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'UserStatusConnection'; ofType: null; }; } }; 'membersCanForkPrivateRepositories': { name: 'membersCanForkPrivateRepositories'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'membersWithRole': { name: 'membersWithRole'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'OrganizationMemberConnection'; ofType: null; }; } }; 'monthlyEstimatedSponsorsIncomeInCents': { name: 'monthlyEstimatedSponsorsIncomeInCents'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Int'; ofType: null; }; } }; 'name': { name: 'name'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'newTeamResourcePath': { name: 'newTeamResourcePath'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'URI'; ofType: null; }; } }; 'newTeamUrl': { name: 'newTeamUrl'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'URI'; ofType: null; }; } }; 'notificationDeliveryRestrictionEnabledSetting': { name: 'notificationDeliveryRestrictionEnabledSetting'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'ENUM'; name: 'NotificationRestrictionSettingValue'; ofType: null; }; } }; 'organizationBillingEmail': { name: 'organizationBillingEmail'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'packages': { name: 'packages'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'PackageConnection'; ofType: null; }; } }; 'pendingMembers': { name: 'pendingMembers'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'UserConnection'; ofType: null; }; } }; 'pinnableItems': { name: 'pinnableItems'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'PinnableItemConnection'; ofType: null; }; } }; 'pinnedItems': { name: 'pinnedItems'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'PinnableItemConnection'; ofType: null; }; } }; 'pinnedItemsRemaining': { name: 'pinnedItemsRemaining'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Int'; ofType: null; }; } }; 'project': { name: 'project'; type: { kind: 'OBJECT'; name: 'Project'; ofType: null; } }; 'projectV2': { name: 'projectV2'; type: { kind: 'OBJECT'; name: 'ProjectV2'; ofType: null; } }; 'projects': { name: 'projects'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'ProjectConnection'; ofType: null; }; } }; 'projectsResourcePath': { name: 'projectsResourcePath'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'URI'; ofType: null; }; } }; 'projectsUrl': { name: 'projectsUrl'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'URI'; ofType: null; }; } }; 'projectsV2': { name: 'projectsV2'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'ProjectV2Connection'; ofType: null; }; } }; 'recentProjects': { name: 'recentProjects'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'ProjectV2Connection'; ofType: null; }; } }; 'repositories': { name: 'repositories'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'RepositoryConnection'; ofType: null; }; } }; 'repository': { name: 'repository'; type: { kind: 'OBJECT'; name: 'Repository'; ofType: null; } }; 'repositoryDiscussionComments': { name: 'repositoryDiscussionComments'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'DiscussionCommentConnection'; ofType: null; }; } }; 'repositoryDiscussions': { name: 'repositoryDiscussions'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'DiscussionConnection'; ofType: null; }; } }; 'repositoryMigrations': { name: 'repositoryMigrations'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'RepositoryMigrationConnection'; ofType: null; }; } }; 'requiresTwoFactorAuthentication': { name: 'requiresTwoFactorAuthentication'; type: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; } }; 'resourcePath': { name: 'resourcePath'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'URI'; ofType: null; }; } }; 'ruleset': { name: 'ruleset'; type: { kind: 'OBJECT'; name: 'RepositoryRuleset'; ofType: null; } }; 'rulesets': { name: 'rulesets'; type: { kind: 'OBJECT'; name: 'RepositoryRulesetConnection'; ofType: null; } }; 'samlIdentityProvider': { name: 'samlIdentityProvider'; type: { kind: 'OBJECT'; name: 'OrganizationIdentityProvider'; ofType: null; } }; 'sponsoring': { name: 'sponsoring'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'SponsorConnection'; ofType: null; }; } }; 'sponsors': { name: 'sponsors'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'SponsorConnection'; ofType: null; }; } }; 'sponsorsActivities': { name: 'sponsorsActivities'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'SponsorsActivityConnection'; ofType: null; }; } }; 'sponsorsListing': { name: 'sponsorsListing'; type: { kind: 'OBJECT'; name: 'SponsorsListing'; ofType: null; } }; 'sponsorshipForViewerAsSponsor': { name: 'sponsorshipForViewerAsSponsor'; type: { kind: 'OBJECT'; name: 'Sponsorship'; ofType: null; } }; 'sponsorshipForViewerAsSponsorable': { name: 'sponsorshipForViewerAsSponsorable'; type: { kind: 'OBJECT'; name: 'Sponsorship'; ofType: null; } }; 'sponsorshipNewsletters': { name: 'sponsorshipNewsletters'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'SponsorshipNewsletterConnection'; ofType: null; }; } }; 'sponsorshipsAsMaintainer': { name: 'sponsorshipsAsMaintainer'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'SponsorshipConnection'; ofType: null; }; } }; 'sponsorshipsAsSponsor': { name: 'sponsorshipsAsSponsor'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'SponsorshipConnection'; ofType: null; }; } }; 'team': { name: 'team'; type: { kind: 'OBJECT'; name: 'Team'; ofType: null; } }; 'teams': { name: 'teams'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'TeamConnection'; ofType: null; }; } }; 'teamsResourcePath': { name: 'teamsResourcePath'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'URI'; ofType: null; }; } }; 'teamsUrl': { name: 'teamsUrl'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'URI'; ofType: null; }; } }; 'totalSponsorshipAmountAsSponsorInCents': { name: 'totalSponsorshipAmountAsSponsorInCents'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'twitterUsername': { name: 'twitterUsername'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'updatedAt': { name: 'updatedAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; 'url': { name: 'url'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'URI'; ofType: null; }; } }; 'viewerCanAdminister': { name: 'viewerCanAdminister'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'viewerCanChangePinnedItems': { name: 'viewerCanChangePinnedItems'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'viewerCanCreateProjects': { name: 'viewerCanCreateProjects'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'viewerCanCreateRepositories': { name: 'viewerCanCreateRepositories'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'viewerCanCreateTeams': { name: 'viewerCanCreateTeams'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'viewerCanSponsor': { name: 'viewerCanSponsor'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'viewerIsAMember': { name: 'viewerIsAMember'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'viewerIsFollowing': { name: 'viewerIsFollowing'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'viewerIsSponsoring': { name: 'viewerIsSponsoring'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'webCommitSignoffRequired': { name: 'webCommitSignoffRequired'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'websiteUrl': { name: 'websiteUrl'; type: { kind: 'SCALAR'; name: 'URI'; ofType: null; } }; }; };
     'OrganizationAuditEntry': { kind: 'UNION'; name: 'OrganizationAuditEntry'; fields: {}; possibleTypes: 'MembersCanDeleteReposClearAuditEntry' | 'MembersCanDeleteReposDisableAuditEntry' | 'MembersCanDeleteReposEnableAuditEntry' | 'OauthApplicationCreateAuditEntry' | 'OrgAddBillingManagerAuditEntry' | 'OrgAddMemberAuditEntry' | 'OrgBlockUserAuditEntry' | 'OrgConfigDisableCollaboratorsOnlyAuditEntry' | 'OrgConfigEnableCollaboratorsOnlyAuditEntry' | 'OrgCreateAuditEntry' | 'OrgDisableOauthAppRestrictionsAuditEntry' | 'OrgDisableSamlAuditEntry' | 'OrgDisableTwoFactorRequirementAuditEntry' | 'OrgEnableOauthAppRestrictionsAuditEntry' | 'OrgEnableSamlAuditEntry' | 'OrgEnableTwoFactorRequirementAuditEntry' | 'OrgInviteMemberAuditEntry' | 'OrgInviteToBusinessAuditEntry' | 'OrgOauthAppAccessApprovedAuditEntry' | 'OrgOauthAppAccessBlockedAuditEntry' | 'OrgOauthAppAccessDeniedAuditEntry' | 'OrgOauthAppAccessRequestedAuditEntry' | 'OrgOauthAppAccessUnblockedAuditEntry' | 'OrgRemoveBillingManagerAuditEntry' | 'OrgRemoveMemberAuditEntry' | 'OrgRemoveOutsideCollaboratorAuditEntry' | 'OrgRestoreMemberAuditEntry' | 'OrgUnblockUserAuditEntry' | 'OrgUpdateDefaultRepositoryPermissionAuditEntry' | 'OrgUpdateMemberAuditEntry' | 'OrgUpdateMemberRepositoryCreationPermissionAuditEntry' | 'OrgUpdateMemberRepositoryInvitationPermissionAuditEntry' | 'PrivateRepositoryForkingDisableAuditEntry' | 'PrivateRepositoryForkingEnableAuditEntry' | 'RepoAccessAuditEntry' | 'RepoAddMemberAuditEntry' | 'RepoAddTopicAuditEntry' | 'RepoArchivedAuditEntry' | 'RepoChangeMergeSettingAuditEntry' | 'RepoConfigDisableAnonymousGitAccessAuditEntry' | 'RepoConfigDisableCollaboratorsOnlyAuditEntry' | 'RepoConfigDisableContributorsOnlyAuditEntry' | 'RepoConfigDisableSockpuppetDisallowedAuditEntry' | 'RepoConfigEnableAnonymousGitAccessAuditEntry' | 'RepoConfigEnableCollaboratorsOnlyAuditEntry' | 'RepoConfigEnableContributorsOnlyAuditEntry' | 'RepoConfigEnableSockpuppetDisallowedAuditEntry' | 'RepoConfigLockAnonymousGitAccessAuditEntry' | 'RepoConfigUnlockAnonymousGitAccessAuditEntry' | 'RepoCreateAuditEntry' | 'RepoDestroyAuditEntry' | 'RepoRemoveMemberAuditEntry' | 'RepoRemoveTopicAuditEntry' | 'RepositoryVisibilityChangeDisableAuditEntry' | 'RepositoryVisibilityChangeEnableAuditEntry' | 'TeamAddMemberAuditEntry' | 'TeamAddRepositoryAuditEntry' | 'TeamChangeParentTeamAuditEntry' | 'TeamRemoveMemberAuditEntry' | 'TeamRemoveRepositoryAuditEntry'; };
