+++ packages/app/src/hooks/useExtendedQuery.ts
@@ -5,7 +5,7 @@ import type { UserProfile } from "../api/auth";
 import { client } from "../api/client";
 import { ContributionQuery } from "../api/query";
 import type { ContributionWeeks } from "../api/types";
-import { useParametersStore } from "../stores";
+import { useParametersStore } from "../stores/parameters";
 
 interface ExtendedQueryProps {
 	name?: string;
