--- packages/app/src/stores/index.ts
@@ -1,3 +1,0 @@
-export * from "./parameters";
-export * from "./scene";
-export * from "./fonts";
